调整三极管测量的算法和返回值，便于显示

+++ main/transistor.hpp
@@ -7,13 +7,12 @@
 class transistorMeasuring_class
 {
 private:
-    
     ADC_class ADC; 
 
 public:
     transistorMeasuring_class(/* args */);
     ~transistorMeasuring_class();
-    int beta_compute(measuringMode_t transistorType);
+    float beta_compute(measuringMode_t transistorType);
 };
